# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "中医药知识库 API 服务 —— 方剂、单味药、对药、论文、分子信息、名家经验管理",
        "title": "AI Medical Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/decoctions": {
            "get": {
                "description": "分页查询方剂，支持关键词搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decoctions"
                ],
                "summary": "获取方剂列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/medicalagent_internal_model.DecoctionBasic"
                                    }
                                },
                                "page": {
                                    "type": "integer"
                                },
                                "page_size": {
                                    "type": "integer"
                                },
                                "total": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员新增方剂记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decoctions"
                ],
                "summary": "创建方剂",
                "parameters": [
                    {
                        "description": "方剂信息",
                        "name": "decoction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medicalagent_internal_model.DecoctionBasic"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/medicalagent_internal_model.DecoctionBasic"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/decoctions/{id}": {
            "get": {
                "description": "根据 ID 获取方剂详细信息，包括组成和主治",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decoctions"
                ],
                "summary": "获取方剂详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "方剂 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/medicalagent_internal_service.DecoctionDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员更新现有方剂记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decoctions"
                ],
                "summary": "更新方剂",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "方剂 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "方剂信息",
                        "name": "decoction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medicalagent_internal_model.DecoctionBasic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员删除指定方剂记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decoctions"
                ],
                "summary": "删除方剂",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "方剂 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "medicalagent_internal_model.DecoctionBasic": {
            "type": "object",
            "properties": {
                "adr": {
                    "type": "string"
                },
                "basis_for_usage_and_dosage": {
                    "type": "string"
                },
                "clinical_suggestion": {
                    "type": "string"
                },
                "clinical_suggestion_basis": {
                    "type": "string"
                },
                "conclusion_of_related_studies": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "crowd_taboo": {
                    "type": "string"
                },
                "decoction_name": {
                    "type": "string"
                },
                "decoction_name_pinyin": {
                    "type": "string"
                },
                "dosage_form": {
                    "type": "string"
                },
                "functionality": {
                    "type": "string"
                },
                "functionality_basis": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link_to_clinical_suggestion": {
                    "type": "string"
                },
                "link_to_functionality_basis": {
                    "type": "string"
                },
                "link_to_related_studies": {
                    "type": "string"
                },
                "link_to_usage_and_dosage": {
                    "type": "string"
                },
                "pathological_examination": {
                    "type": "string"
                },
                "related_studies": {
                    "type": "string"
                },
                "related_toxic_herbs": {
                    "type": "string"
                },
                "symptom_contraindications": {
                    "type": "string"
                },
                "toxicity_mechanism": {
                    "type": "string"
                },
                "typical_cases_of_adr": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_and_dosage": {
                    "type": "string"
                },
                "virulence": {
                    "type": "string"
                }
            }
        },
        "medicalagent_internal_model.DecoctionToxicCompound": {
            "type": "object",
            "properties": {
                "cas": {
                    "type": "string"
                },
                "compound_name": {
                    "type": "string"
                },
                "compound_type": {
                    "type": "string"
                },
                "decoction_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "molecular_formula": {
                    "type": "string"
                },
                "record_number": {
                    "type": "integer"
                }
            }
        },
...
        "medicalagent_internal_service.DecoctionDetail": {
            "type": "object",
            "properties": {
                "adr": {
                    "type": "string"
                },
                "basis_for_usage_and_dosage": {
                    "type": "string"
                },
                "clinical_suggestion": {
                    "type": "string"
                },
                "clinical_suggestion_basis": {
                    "type": "string"
                },
                "conclusion_of_related_studies": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "crowd_taboo": {
                    "type": "string"
                },
                "decoction_name": {
                    "type": "string"
                },
                "decoction_name_pinyin": {
                    "type": "string"
                },
                "dosage_form": {
                    "type": "string"
                },
                "functionality": {
                    "type": "string"
                },
                "functionality_basis": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link_to_clinical_suggestion": {
                    "type": "string"
                },
                "link_to_functionality_basis": {
                    "type": "string"
                },
                "link_to_related_studies": {
                    "type": "string"
                },
                "link_to_usage_and_dosage": {
                    "type": "string"
                },
                "pathological_examination": {
                    "type": "string"
                },
                "related_studies": {
                    "type": "string"
                },
                "related_toxic_herbs": {
                    "type": "string"
                },
                "symptom_contraindications": {
                    "type": "string"
                },
                "toxic_compounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/medicalagent_internal_model.DecoctionToxicCompound"
                    }
                },
                "toxicity_mechanism": {
                    "type": "string"
                },
                "typical_cases_of_adr": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_and_dosage": {
                    "type": "string"
                },
                "virulence": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AgentTokenAuth": {
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}